Before changing sytem with unique and mode
Before we collect anntations and check later which are >=2 in sum.

Now we are creating a correct annotation row for each annotator and then
look first how mnay single values are there (unique) and if less than 3
we use mode to find the most common one. unique is 3 we either choose
Renees or Nadice annotations. Maybe first Renees, if renees is Nan then
we use nadice once